requirements.txt
absl-py==1.0.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==22.3.0
bleach==4.1.0
cachetools==4.2.4
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
configargparse==1.7
contextvars==2.4
cycler==0.11.0
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
einops==0.3.2
entrypoints==0.4
filelock==3.4.1
freetype-py==2.2.0
frozenlist==1.2.0
fsspec==2022.1.0
future==0.18.2
gdown==4.7.1
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.45.0
huggingface-hub==0.4.0
hydra-slayer==0.4.0
idna==3.3
idna-ssl==1.1.0
imageio==2.9.0
imageio-ffmpeg==0.4.2
immutables==0.17
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipdb==0.13.9
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.17.2
jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.0.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyter-packaging==0.10.6
jupyter-server==1.13.1
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.1.0
keopscore==2.1
kiwisolver==1.3.1
kornia==0.5.1
llvmlite==0.35.0
lpips==0.1.4
markdown==3.3.6
markupsafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
multidict==5.2.0
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
networkx==2.5.1
notebook==6.4.10
numba==0.52.0
oauthlib==3.2.0
omegaconf==2.1.1
open3d==0.9.0.0
opencv-python==4.5.3.56
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.4.0
platformdirs==2.4.0
plyfile==0.7.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==4.21.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.0
pycparser==2.21
pydeprecate==0.3.1
pygifsicle==1.0.7
pyglet==1.5.23
pygments==2.12.0
pykeops==2.1
pymcubes==0.1.2
pyopengl==3.1.0
pyparsing==3.0.8
pyquaternion==0.9.9
pyrender==0.1.45
pyrsistent==0.18.0
pysocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.5.4
pytorch-ranger==0.1.1
pytz==2022.1
pywavelets==1.1.1
pyyaml==6.0
pyzmq==22.3.0
qtconsole==5.2.2
qtpy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.16.1
scikit-learn==0.24.2
scikit-video==1.1.11
scipy==1.5.4
send2trash==1.8.0
sniffio==1.2.0
soupsieve==2.3.2.post1
support-developer==1.0.5
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardx==2.2
terminado==0.13.0
testpath==0.6.0
threadpoolctl==3.1.0
timm==0.6.11
toml==0.10.2
tomli==1.2.3
tomlkit==0.10.2
torch-efficient-distloss==0.1.3
torch-optimizer==0.3.0
torchmetrics==0.8.1
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
trimesh==3.11.2
typed-ast==1.5.3
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
werkzeug==2.0.3
widgetsnbextension==3.6.0
yarl==1.7.2
zipp==3.6.0